+++ CHANGELOG.md
@@ -5,17 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### New
+* configparser template to serialize ini/properties files
 
 ## [1.2.0] - 2021-12-03
-### Added
+### New
 - Support for repeated sections
 
 ## [1.1.0] - 2021-02-24
-### Added
+### New
 - sysctl: serialize boolean to lower-case
 
 ## 1.0.0 - 2021-02-22
-### Added
+### New
 - Initial release
 
 [Unreleased]: https://github.com/jgraichen/salt-template/compare/v1.2.0...HEAD